requirements.txt
# # Core Dependencies
# streamlit==1.31.0
# python-dotenv==1.0.0

# # Authentication
# clerk-backend-api==1.0.0

# # Database - Updated for Python 3.13
# sqlalchemy==2.0.27
# psycopg2-binary==2.9.9
# alembic==1.13.1
# pgvector==0.2.5

# # AWS Services
# boto3==1.34.51
# botocore==1.34.51

# # Encryption
# # cryptography==42.0.5
# cryptography==43.0.3


# # PDF Processing
# PyPDF2==3.0.1
# pdfplumber==0.11.0
# pypdf==4.0.2

# # Text Processing
# langchain==0.1.9
# langchain-community==0.0.24
# tiktoken==0.6.0

# # Email
# python-multipart==0.0.9

# # Validation - Updated for Python 3.13
# # pydantic==2.6.3
# pydantic>=2.8.2,<2.9.0
# pydantic-settings==2.2.1

# # Utilities
# numpy==1.26.4
# pandas==2.2.1
# requests==2.31.0

# # Testing
# pytest==8.0.2
# pytest-asyncio==0.23.5
# faker==23.2.1

# # Monitoring & Logging
# python-json-logger==2.0.7

# # Additional dependencies for Python 3.13 compatibility
# typing-extensions==4.10.0

# # Added for Clerk JWT verification and API server
# python-jose[cryptography]==3.3.0
# fastapi==0.100.0
# uvicorn[standard]==0.22.0


# Core Dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0

# Authentication
clerk-backend-api>=1.0.0

# Database
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9
alembic>=1.13.1
pgvector>=0.2.5

# AWS Services
boto3>=1.34.51

# Encryption
cryptography>=43.0.3

# PDF Processing
PyPDF2>=3.0.1
pdfplumber>=0.11.0
pypdf>=4.0.2

# Text Processing
langchain>=0.1.9
langchain-community>=0.0.24
tiktoken>=0.6.0

# Email
python-multipart>=0.0.9

# Validation
pydantic>=2.8.2
pydantic-settings>=2.2.1

# Utilities
numpy>=2.0.0
pandas>=2.2.1
requests>=2.31.0

# Monitoring & Logging
python-json-logger>=2.0.7

# Type hints
typing-extensions>=4.10.0

# JWT verification
PyJWT[crypto]>=2.8.0
python-jose[cryptography]>=3.3.0

# API server
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Testing
pytest>=8.0.2
pytest-asyncio>=0.23.5
faker>=23.2.1